Move WellContainer to new loci.formats.internal package
